+++ src/ops/ops.ts
@@ -16,6 +16,7 @@
  */
 
 export * from './batchnorm';
+export * from './complex_ops';
 export * from './concat';
 export * from './conv';
 export * from './matmul';
